refactor(firestoreBackup): move backup related functions to firebase-utils
